partial login updates
