+++ spotify_colors/src/App.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 // import './App.css';
 import PageRouter from "./page_router.tsx";
+// import Connection from "./components/connection.tsx";
 
 function App() {
   return (
