+++ spotify_colors/src/pages/no_page.tsx
@@ -1,0 +1,9 @@
+import React from "react"
+
+function NoPage() {
+    return (
+        <h1>"There's nothing here"</h1>
+    );
+}
+
+export default NoPage;
