+++ spotify_colors/src/pages/redirect.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Redirect() {
+  return (
+    <div>
+      <h1>redirect page</h1>
+    </div>
+  );
+}
+
+export default Redirect;
